modify according to commend
